requirements.txt
amqp==2.5.2
asn1crypto==1.3.0
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
configparser==4.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cx-Oracle==7.3
Django==3.0.3
django-bootstrap3==12.0.3
django-celery-beat==2.0.0
django-celery-results==1.2.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-swagger==2.2.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
importlib-metadata==1.5.0
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.7
MarkupSafe==1.1.1
more-itertools==8.2.0
mysqlclient==1.4.6
numpy==1.18.0
openapi-codec==1.3.2
paramiko==2.7.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.3.0
tornado==6.0.3
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
zipp==3.0.0
channels==2.4.0